% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_counties.R
\name{add_counties}
\alias{add_counties}
\title{Add county, FIPS, and geometry to data frame with zip codes}
\usage{
add_counties(df, state, zip, fips = FALSE, geo = FALSE, sf = FALSE)
}
\arguments{
\item{df}{data frame}

\item{state}{bare column name column containing state abbreviations}

\item{zip}{bare column name containing zip codes}

\item{fips}{add county FIPS code column, default is \code{FALSE}}

\item{geo}{add county geometry column, default is \code{FALSE}}

\item{sf}{convert to an \code{sf} object, default is \code{FALSE}}
}
\description{
Add county, FIPS, and geometry to data frame with zip codes
}
\examples{
(ex <- dplyr::tibble(state = "GA", zip = "31605"))

# Add county and latitude/longitude
ex |> add_counties(state, zip)

# Add county FIPS
ex |> add_counties(state, zip, fips = TRUE)

# Add county `geometry` column,
# based on county FIPS column
ex |> add_counties(state, zip, fips = TRUE, geo = TRUE)

# Convert to an `sf` object
ex |> add_counties(state, zip, fips = TRUE, geo = TRUE, sf = TRUE)

}
