% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ins.R
\name{add_ifelse}
\alias{add_ifelse}
\title{Search in data frame column if search term is not \code{NULL}}
\usage{
add_ifelse(x, df, dfcol, by)
}
\arguments{
\item{x}{A \verb{<character>} or \verb{<symbol>} specifying the search term}

\item{df}{A \verb{<data.frame>} or \verb{<tibble>}}

\item{dfcol}{A \verb{<character>} or \verb{<symbol>} specifying the column to search in}

\item{by}{A \verb{<character>} or \verb{<symbol>} specifying the column to nest by}
}
\value{
A \verb{<data.frame>} or \verb{<tibble>}
}
\description{
Search in data frame column if search term is not \code{NULL}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

df <- northstar::get_mue_edits()

add_ifelse(x = "Practitioner",
           df = df,
           dfcol = df$mue_service_type,
           by = mue_service_type)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
