% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_ops.R
\name{arrows}
\alias{arrows}
\title{Wrapper for \code{\link[=paste0]{paste0()}} that adds angle brackets}
\usage{
arrows(x)
}
\arguments{
\item{x}{\verb{<chr>} string}
}
\description{
Wrapper for \code{\link[=paste0]{paste0()}} that adds angle brackets
}
\keyword{internal}
