% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{as_chr}
\alias{as_chr}
\title{Coerce vector to \verb{<chr>}}
\usage{
as_chr(x)
}
\arguments{
\item{x}{vector}
}
\value{
\verb{<chr>} vector
}
\description{
Coerce vector to \verb{<chr>}
}
\keyword{internal}
