% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{as_date}
\alias{as_date}
\title{Coerce vector to \verb{<date>} class}
\usage{
as_date(x, ..., fmt = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{vector}

\item{...}{additional arguments}

\item{fmt}{\verb{<chr>} format; default is \code{"\%Y-\%m-\%d"}}
}
\value{
\verb{<date>} vector
}
\description{
Coerce vector to \verb{<date>} class
}
\keyword{internal}
