% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{as_int}
\alias{as_int}
\title{Coerce vector to \verb{<int>}}
\usage{
as_int(x)
}
\arguments{
\item{x}{vector}
}
\value{
\verb{<int>} vector
}
\description{
Coerce vector to \verb{<int>}
}
\keyword{internal}
