% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{as_num}
\alias{as_num}
\title{Coerce vector to \verb{<num>} class}
\usage{
as_num(x)
}
\arguments{
\item{x}{vector}
}
\value{
\verb{<num>} vector
}
\description{
Coerce vector to \verb{<num>} class
}
\keyword{internal}
