% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dar.R
\name{avg_dar}
\alias{avg_dar}
\title{Calculate Average Days in AR}
\usage{
avg_dar(df, date, gct, earb, dart = 35, period = c("month", "quarter"))
}
\arguments{
\item{df}{\verb{<data.frame>} containing at least three columns: a date column, a
gross charges column and an ending AR column}

\item{date}{column containing a date within the month that Days in AR is to
be calculated}

\item{gct}{column containing a month's total Gross Charges}

\item{earb}{column containing a month's Ending AR balance}

\item{dart}{Days in AR target, default is \code{35} days}

\item{period}{\verb{<chr>} string specifying the period; either "month" or
"quarter"}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Calculate Average Days in AR
}
\examples{
avg_dar(df     = dar_example(),
        date   = date,
        gct    = gct,
        earb   = earb,
        dart   = 35,
        period = "month")

avg_dar(df     = dar_example(),
        date   = date,
        gct    = gct,
        earb   = earb,
        dart   = 35,
        period = "quarter")

}
