% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_data.R
\name{check_chars_numeric}
\alias{check_chars_numeric}
\title{Check that \code{x} contains numbers only}
\usage{
check_chars_numeric(x)
}
\arguments{
\item{x}{vector}
}
\description{
Check that \code{x} contains numbers only
}
\keyword{internal}
