% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_class.R
\name{check_class}
\alias{check_class}
\title{Check Class of Argument Inputs}
\usage{
check_class(
  ...,
  .expected_class = c("numeric", "character"),
  .call = rlang::caller_env()
)
}
\arguments{
\item{...}{Objects to be checked for class.}

\item{.expected_class}{Character. The name of the class against which objects
should be checked.}

\item{.call}{The environment in which this function is to be called.}
}
\description{
Check Class of Argument Inputs
}
\keyword{internal}
