% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_data.R, R/random_npi_generator.R
\name{check_first_char}
\alias{check_first_char}
\title{Check that \code{x} begin with 1 or 2}
\usage{
check_first_char(x)

check_first_char(x)
}
\arguments{
\item{x}{vector}
}
\description{
Check that \code{x} begin with 1 or 2

Check that \code{x} begin with 1 or 2
}
\keyword{internal}
