% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_data.R, R/random_npi_generator.R
\name{check_length}
\alias{check_length}
\title{Check length of \code{x} is 1}
\usage{
check_length(x)

check_length(x)
}
\arguments{
\item{x}{vector}
}
\description{
Check length of \code{x} is 1

Check length of \code{x} is 1
}
\keyword{internal}
