% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_data.R, R/random_npi_generator.R
\name{check_nchars_10}
\alias{check_nchars_10}
\title{Check that \code{x} is 10 characters long}
\usage{
check_nchars_10(x)

check_nchars_10(x)
}
\arguments{
\item{x}{vector}
}
\description{
Check that \code{x} is 10 characters long

Check that \code{x} is 10 characters long
}
\keyword{internal}
