% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{chg_abs}
\alias{chg_abs}
\title{Lagged absolute change}
\usage{
chg_abs(x, n = 1L, fill_na = 0L)
}
\arguments{
\item{x}{numeric vector}

\item{n}{values to offset}

\item{fill_na}{fill value for any NAs; default is 0}
}
\description{
Lagged absolute change
}
\examples{
mock_provider(2020:2025) |>
  dplyr::mutate(change = chg_abs(payment), .by = group)

}
