% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chchchchanges.R
\name{chg_pct}
\alias{chg_pct}
\title{Lagged percentage change}
\usage{
chg_pct(x, n = 1L, fill_na = 0L)
}
\arguments{
\item{x}{numeric vector}

\item{n}{values to offset}

\item{fill_na}{fill value for any NAs; default is 0}
}
\description{
Lagged percentage change
}
\examples{
fuimus:::provider_data(2020:2025) |>
  dplyr::mutate(change = chg_pct(net_payment),
  .by = group)

}
