% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_number.R
\name{clean_number}
\alias{clean_number}
\title{Clean character vector of numbers}
\usage{
clean_number(x)
}
\arguments{
\item{x}{\verb{<chr>} vector of numbers}
}
\value{
a \verb{<dbl>} vector of numbers
}
\description{
Clean character vector of numbers
}
\examples{
clean_number(c("20\%", "21,125,458", "$123"))

}
