% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{collapser}
\alias{collapser}
\title{Wrapper for \code{\link[=paste0]{paste0()}} that collapses result}
\usage{
collapser(x)
}
\arguments{
\item{x}{A split \verb{<chr>} vector}
}
\value{
A collapsed \verb{<chr>} string
}
\description{
Wrapper for \code{\link[=paste0]{paste0()}} that collapses result
}
\examples{
collapser(c("X", "Y", "Z"))

}
