% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colon.R
\name{colon}
\alias{colon}
\title{Number Sequence Starting at 1}
\usage{
colon(n)
}
\arguments{
\item{n}{\verb{<integer>} Ending number of sequence}
}
\value{
\verb{<integer>} vector of numbers from 1 to \code{n}
}
\description{
Number Sequence Starting at 1
}
\examples{
colon(50)

colon(-20)

colon(0 + 150 - 145)

}
