% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_gt.R
\name{color_text}
\alias{color_text}
\title{Color text using HTML color styling}
\usage{
color_text(text, hex_code)
}
\arguments{
\item{text}{\verb{<character>} string of text to}

\item{hex_code}{\verb{<character>} string, color hex code}
}
\value{
\verb{<character>} string
}
\description{
Useful in conjuction with ggtext and glue
}
\examples{
color_text("Hello there!", "#BD43BF")

}
