% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{columns_to_character}
\alias{columns_to_character}
\title{Coerce all columns to type character}
\usage{
columns_to_character(data)
}
\arguments{
\item{data}{\verb{<data.frame>}}
}
\value{
\verb{<data.frame>} with columns coerced to character
}
\description{
Coerce all columns to type character
}
\examples{
(x <- dplyr::tibble(int = c(1:9, NA_integer_), chr = letters[1:10], date = Sys.Date()))

columns_to_character(x)

}
\seealso{
Other tidyverse: 
\code{\link{gchop}()},
\code{\link{glue_chr}()},
\code{\link{glue_data_chr}()},
\code{\link{named_group_split}()},
\code{\link{search_for}()}
}
\concept{tidyverse}
\keyword{helpers}
