% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_missing.R
\name{count_missing}
\alias{count_missing}
\title{Count missing values}
\usage{
count_missing(df, group_vars, x_var)
}
\arguments{
\item{df}{A \verb{<data.frame>} or \verb{<tibble>}}

\item{group_vars}{A \verb{<character>} or \verb{<symbol>} vector of the variables to
group by}

\item{x_var}{A \verb{<character>} or \verb{<symbol>}  vector of the variable to count
missing values for}
}
\description{
Count missing values
}
\examples{
dplyr::tibble(x = 1:10,
              y = 1:10,
              z = letters[1:10]) |>
              count_missing(z, x)
}
