% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_prop.R
\name{count_prop}
\alias{count_prop}
\title{Count and calculate proportion of each group}
\usage{
count_prop(df, var, sort = FALSE)
}
\arguments{
\item{df}{A \verb{<data.frame>} or \verb{<tibble>}}

\item{var}{A \verb{<character>} or \verb{<symbol>} specifying the column to count}

\item{sort}{A \verb{<logical>} indicating whether to sort the output by frequency,
default is \code{FALSE}}
}
\description{
Count and calculate proportion of each group
}
\examples{
dplyr::tibble(x = 1:10,
              y = sample(1:1000, length(x))) |>
              count_prop(x)

}
