% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{count_prop_multi}
\alias{count_prop_multi}
\title{Count and calculate proportion of each group}
\usage{
count_prop_multi(df, rows, cols, sort = FALSE, na.rm = FALSE)
}
\arguments{
\item{df}{A \verb{<data.frame>} or \verb{<tibble>}}

\item{rows}{A \verb{<character>} or \verb{<symbol>} specifying the rows to count}

\item{cols}{A \verb{<character>} or \verb{<symbol>} specifying the columns to count}

\item{sort}{A \verb{<logical>} indicating whether to sort the output by frequency,
default is \code{FALSE}}

\item{na.rm}{A \verb{<logical>} indicating whether to remove missing values from
the count, default is \code{FALSE}}
}
\description{
Count and calculate proportion of each group
}
\examples{
ggplot2::diamonds |> head()

ggplot2::diamonds |>
  count_prop_multi(
  c(clarity, color),
  cut,
  sort = TRUE)

}
