% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{count_wide}
\alias{count_wide}
\title{Count rows and columns and pivot to wide format}
\usage{
count_wide(df, rows, cols)
}
\arguments{
\item{df}{A \verb{<data.frame>} or \verb{<tibble>}}

\item{rows}{A \verb{<character>} or \verb{<symbol>} specifying the rows to count}

\item{cols}{A \verb{<character>} or \verb{<symbol>} specifying the columns to count}
}
\description{
Count rows and columns and pivot to wide format
}
\examples{
fuimus:::forager_data(10) |>
  count_wide(
    c(ins_class),
    payer)

}
