% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_ops.R
\name{create_vec}
\alias{create_vec}
\title{Generate code from a vector of values}
\usage{
create_vec(x, collapse = ", ", enclose = c("x = c(", ")"), style = TRUE, ...)
}
\arguments{
\item{x}{\verb{<vec>} vector}

\item{collapse}{\verb{<chr>} Separator between sequences, default \code{", "}}

\item{enclose}{\verb{<chr>} \emph{(optional)} Vector of \code{length(x) == 2} with which to enclose output}

\item{style}{\verb{<lgl>} Apply \code{\link[styler:style_text]{styler::style_text()}} to output, default \code{TRUE}}

\item{...}{Additional arguments passed to \code{\link[styler:tidyverse_style]{styler::tidyverse_style()}}}
}
\value{
\verb{<chr>} vector of valid R code
}
\description{
Generate code from a vector of values
}
\examples{
hcpcs <- c(
   "63091", "81536", "99305", "63012",
   "78835", "E0185", "44408", "87485",
   "36015", "0581F", "33478", "21184",
   "42999", "15155", "76705", "23412",
   "99406", "0585T", "0272T", "92507")

create_vec(x = hcpcs)

}
