% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dar.R
\name{dar_quarter}
\alias{dar_quarter}
\title{Calculate Quarterly Days in AR}
\usage{
dar_quarter(df, date, gct, earb, dart = 35)
}
\arguments{
\item{df}{\verb{<data.frame>} containing at least three columns: a date column, a
gross charges column and an ending AR column}

\item{date}{column containing a date within the month that Days in AR is to
be calculated}

\item{gct}{column containing a month's total Gross Charges}

\item{earb}{column containing a month's Ending AR balance}

\item{dart}{Days in AR target, default is \code{35} days}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Calculate Quarterly Days in AR
}
\examples{
dar_quarter(
   dar_example(),
   date,
   gct,
   earb,
   dart = 40)

}
