% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{data_types}
\alias{data_types}
\title{Return data.frame of column types}
\usage{
data_types(data)
}
\arguments{
\item{data}{\verb{<data.frame>}}
}
\value{
\verb{<data.frame>} or \verb{<tibble>}
}
\description{
Return data.frame of column types
}
\examples{
(x <- dplyr::tibble(int = c(1:9, NA_integer_), chr = letters[1:10], date = Sys.Date()))

data_types(x)

}
