% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{delist}
\alias{delist}
\title{Unlist with no names}
\usage{
delist(x)
}
\arguments{
\item{x}{Named or unnamed \verb{<list>}}
}
\value{
Unnamed \verb{<character>} vector
}
\description{
Unlist with no names
}
\examples{
delist(list(x = "XYZ"))

delist(list("XYZ"))

}
\seealso{
Other base: 
\code{\link{as_character}()},
\code{\link{colon}()},
\code{\link{desplit}()},
\code{\link{interpolate}()},
\code{\link{invert_named}()},
\code{\link{not_null}()},
\code{\link{null}()}
}
\concept{base}
\keyword{helpers}
