% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delister.R
\name{delister}
\alias{delister}
\title{Wrapper for \code{\link[=unlist]{unlist()}}, with \code{use.names} set to \code{FALSE}}
\usage{
delister(x)
}
\arguments{
\item{x}{A named \verb{<list>}}
}
\value{
An unnamed \verb{<chr>} vector
}
\description{
Wrapper for \code{\link[=unlist]{unlist()}}, with \code{use.names} set to \code{FALSE}
}
\examples{
delister(list(x = "XYZ"))

}
