% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe a dataset}
\usage{
describe(df, ...)
}
\arguments{
\item{df}{\verb{<data.frame>} desc}

\item{...}{\verb{<dots>} tidyselect columns}
}
\value{
\verb{<tibble>} of summary statistics
}
\description{
Describe a dataset
}
\examples{
describe(mock_provider(2000:2020))

describe(mock_forager(200), !dplyr::starts_with("date"))

}
