% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{desplit}
\alias{desplit}
\title{Unlist, unname and split}
\usage{
desplit(x)
}
\arguments{
\item{x}{\verb{<chr>} string or named \verb{<list>}}
}
\description{
Unlist, unname and split
}
\keyword{internal}
