% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{display_long}
\alias{display_long}
\title{Pivot data frame to long format for easy printing}
\usage{
display_long(data)
}
\arguments{
\item{data}{\verb{<data.frame>}}
}
\value{
a \verb{<data.frame>} or \verb{<tibble>} in long format
}
\description{
Pivot data frame to long format for easy printing
}
\examples{
(x <- dplyr::tibble(int = c(1:9, NA_integer_), chr = letters[1:10], date = Sys.Date()))

display_long(x)

}
