% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{empty}
\alias{empty}
\title{Is Vector Empty?}
\usage{
empty(x)
}
\arguments{
\item{x}{vector}
}
\value{
\verb{<lgl>} \code{TRUE} if empty, else \code{FALSE}
}
\description{
Is Vector Empty?
}
\keyword{internal}
