% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{expand_date_range}
\alias{expand_date_range}
\title{Expand date range to include all dates}
\usage{
expand_date_range(x, name)
}
\arguments{
\item{x}{\verb{<date>} vector of dates}

\item{name}{\verb{<chr>} new column name for vector of dates}
}
\value{
\verb{<tbl>} with expanded date range
}
\description{
Expand date range to include all dates
}
\examples{
expand_date_range(
   seq.Date(
   from = as.Date("2020-01-01"),
   to = as.Date("2020-01-01") + 2057,
   by = 1),
   "date"
)
}
