% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{false}
\alias{false}
\title{Is \code{x} \code{FALSE}?}
\usage{
false(x)
}
\arguments{
\item{x}{input}
}
\value{
\verb{<lgl>} \code{TRUE} if \code{x} is \code{FALSE}, else \code{FALSE}
}
\description{
Is \code{x} \code{FALSE}?
}
\keyword{internal}
