% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_misc.R
\name{fancy_ts}
\alias{fancy_ts}
\title{Plots a fancy time series}
\usage{
fancy_ts(df, val, group)
}
\arguments{
\item{df}{data.frame}

\item{val}{value var}

\item{group}{group var}
}
\value{
A \link[tibble:tibble-package]{tibble} with the summarized data
}
\description{
Plots a fancy time series
}
\examples{
df <- dplyr::tibble(
dist1 = sort(rnorm(50, 5, 2)),
dist2 = sort(rnorm(50, 8, 3)),
dist4 = sort(rnorm(50, 15, 1)),
date = seq.Date(
as.Date("2022-01-01"),
as.Date("2022-04-10"),
by = "2 days")
) |>
tidyr::pivot_longer(
cols = -date,
names_to = "dist_name",
values_to = "value")

df

fancy_ts(df, value, dist_name) +
gg_theme()

}
