% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{forager_data}
\alias{forager_data}
\title{Generate mock coding/billing data frame}
\usage{
forager_data(rows = 10, unnest = FALSE)
}
\arguments{
\item{rows}{number of rows to generate; default is 10}

\item{unnest}{a logical indicating whether to unnest the dates column;
default is \code{FALSE}}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Generate mock coding/billing data frame
}
\examples{
forager_data(rows = 5)

}
\seealso{
Other mock: 
\code{\link{provider_data}()},
\code{\link{random_npi_generator}()}
}
\concept{mock}
