% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs.R
\name{gchop}
\alias{gchop}
\title{Chop Vector by Group}
\usage{
gchop(v, g)
}
\arguments{
\item{v}{\verb{<character>} vector}

\item{g}{\verb{<integer>} group}
}
\value{
\verb{<list>}
}
\description{
Chop Vector by Group
}
\examples{
(v <- c("222", "280", "3020", "8690", "G0294", "G8126"))

(g <- sample(1:2, size = length(v), replace = TRUE))

gchop(v, g)

}
\seealso{
Other vctrs: 
\code{\link{empty}()},
\code{\link{null_if_empty}()},
\code{\link{search_in}()},
\code{\link{search_in_impl}()}
}
\concept{vctrs}
