% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{gelm}
\alias{gelm}
\title{\code{getelem} with more flexibility}
\usage{
gelm(l, e, m = "re", ...)
}
\arguments{
\item{l}{named \verb{<list>}}

\item{e}{\verb{<chr>} element name; can be a regex pattern if \code{m} is \code{"re"}}

\item{m}{\verb{<chr>} mode; default is \code{"re"} (regex) or \code{"df"} (data frame)}

\item{...}{additional arguments to pass to \code{collapse::get_elem()}}
}
\value{
\verb{<chr>} or \verb{<data.frame>}, depending on \code{m}
}
\description{
\code{getelem} with more flexibility
}
\keyword{internal}
