% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{geomean}
\alias{geomean}
\title{Calculate Geometric Mean}
\usage{
geomean(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Average \code{\link[=rate_of_return]{rate_of_return()}}
}
\examples{
x <- mock_provider(2020:2025)

x |>
  dplyr::group_by(group) |>
  rate_of_return(payment)

x |>
  dplyr::group_by(group) |>
  rate_of_return(payment) |>
  dplyr::summarise(gmean = geomean(payment_ror), .by = group)

}
