% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{get_pin}
\alias{get_pin}
\title{Get pinned dataset from mount_board()}
\usage{
get_pin(pin, ...)
}
\arguments{
\item{pin}{\verb{<chr>} string name of pinned dataset}

\item{...}{additional arguments passed to mount_board()}
}
\value{
\verb{<tibble>} or \verb{<data.frame>}
}
\description{
Get pinned dataset from mount_board()
}
\keyword{internal}
