% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{getelem}
\alias{getelem}
\title{Get named element from list}
\usage{
getelem(l, e)
}
\arguments{
\item{l}{named \verb{<list>}}

\item{e}{\verb{<chr>} element name; can be a regex pattern}
}
\value{
\verb{<chr>} element value
}
\description{
Get named element from list
}
\keyword{internal}
