% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{gh_raw}
\alias{gh_raw}
\title{Return GitHub raw url}
\usage{
gh_raw(x)
}
\arguments{
\item{x}{\verb{<chr>} string}
}
\value{
\verb{<chr>} GitHub raw url
}
\description{
Return GitHub raw url
}
\examples{
gh_raw("andrewbruce/example/main/inst/pins/")

}
\keyword{internal}
