% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_development.R
\name{git_commit}
\alias{git_commit}
\title{Commit and push changes with gert}
\usage{
git_commit(msg)
}
\arguments{
\item{msg}{\verb{<chr>} commit message}
}
\value{
nothing
}
\description{
Commit and push changes with gert
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
git_commit("* bump dev version")
\dontshow{\}) # examplesIf}
}
\keyword{internal}
