% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_ops.R
\name{glue_chr}
\alias{glue_chr}
\title{Wrapper for \code{as.character(glue(x))}}
\usage{
glue_chr(...)
}
\arguments{
\item{...}{dots to pass to glue function}
}
\value{
\verb{<chr>} vector
}
\description{
Wrapper for \code{as.character(glue(x))}
}
