% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{glue_data_chr}
\alias{glue_data_chr}
\title{Coerce glue_data() output to character}
\usage{
glue_data_chr(.x, ...)
}
\arguments{
\item{.x}{\verb{<vec>} vector to pass to glue_data()}

\item{...}{dots to pass to glue function}
}
\value{
\verb{<chr>} vector
}
\description{
Coerce glue_data() output to character
}
\seealso{
Other tidyverse: 
\code{\link{columns_to_character}()},
\code{\link{gchop}()},
\code{\link{glue_chr}()},
\code{\link{named_group_split}()},
\code{\link{search_for}()}
}
\concept{tidyverse}
\keyword{helpers}
