% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Infix \code{!x \%in\% y}}
\usage{
x \%nin\% y
}
\arguments{
\item{x, y}{vectors}
}
\value{
\code{TRUE} if \code{x} is not in \code{y}, else \code{FALSE}
}
\description{
Infix \code{!x \%in\% y}
}
\examples{
"a" \%nin\% c("a", "b", "c")

}
