% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{\%or\%}
\alias{\%or\%}
\title{Infix \code{!x \%in\% y}}
\usage{
x \%or\% y
}
\arguments{
\item{x, y}{vectors}
}
\value{
\code{y} if \code{x} is length 0, else \code{x}
}
\description{
Infix \code{!x \%in\% y}
}
\examples{
NULL \%or\% 2474701

}
