% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_gt.R
\name{gt_marks}
\alias{gt_marks}
\title{gt Marks}
\usage{
gt_marks(gt_tbl, cols)
}
\arguments{
\item{gt_tbl}{A \verb{<gt_tbl>} object}

\item{cols}{columns to target}
}
\value{
A \verb{<gt_tbl>} object with marks
}
\description{
gt Marks
}
\examples{
dplyr::tibble(pass = c(TRUE, FALSE, TRUE, FALSE)) |>
  gt::gt() |>
  gt_marks(cols = pass)

}
