% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{histo}
\alias{histo}
\title{Inline histogram}
\usage{
histo(x, width = 10)
}
\arguments{
\item{x}{description}

\item{width}{description}
}
\value{
description
}
\description{
Inline histogram
}
\examples{
histo(x = rnorm(100))

histo(x = runif(100))

histo(x = rnorm(100), width = 5)

histo(x = runif(100), width = 5)

}
\keyword{internal}
