% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_misc.R
\name{histogram}
\alias{histogram}
\title{Histogram}
\usage{
histogram(df, var, binwidth)
}
\arguments{
\item{df}{df}

\item{var}{var}

\item{binwidth}{binwidth}
}
\value{
ggplot2 histogram
}
\description{
Histogram
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

ggplot2::diamonds |> histogram(carat, 0.1)
\dontshow{\}) # examplesIf}
}
