% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{if_empty_null}
\alias{if_empty_null}
\title{If \code{x} is empty, \code{NULL}, else \code{x}}
\usage{
if_empty_null(x)
}
\arguments{
\item{x}{vector}
}
\value{
\code{x} or \code{NULL}
}
\description{
If \code{x} is empty, \code{NULL}, else \code{x}
}
\keyword{internal}
