% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastverse.R
\name{iif_else}
\alias{iif_else}
\title{If Else wrapper using \code{\link[kit:iif]{kit::iif()}}}
\usage{
iif_else(x, yes, no)
}
\arguments{
\item{x}{\verb{<logical>} vector}

\item{yes, no}{Values to return depending on TRUE/FALSE element of \code{x}. Must
be same type and be either length 1 or same length of \code{x}.}
}
\value{
vector of same length as \code{x} and attributes as \code{yes}. Data values
are taken from values of \code{yes} and \code{no}.
}
\description{
If Else wrapper using \code{\link[kit:iif]{kit::iif()}}
}
\examples{
x <- c(1:4, 3:2, 1:4)

iif_else(x > 2L, x, x - 1L)

}
