% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{initialize_package}
\alias{initialize_package}
\title{Initialize a new R package}
\usage{
initialize_package(pkgname)
}
\arguments{
\item{pkgname}{\verb{<chr>} name of package}
}
\value{
nothing
}
\description{
Initialize a new R package
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
initialize_package("package")
\dontshow{\}) # examplesIf}
}
\keyword{internal}
