% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{interpolate}
\alias{interpolate}
\title{Linearly interpolate values between two points}
\usage{
interpolate(x, y, method = c("approx", "spline"))
}
\arguments{
\item{x, y}{\verb{<numeric>} vectors giving the coordinates of the points to be
interpolated}

\item{method}{\verb{<character>} interpolation method, default is \code{"approx"}}
}
\value{
\verb{<numeric>} vector of interpolated values
}
\description{
\href{https://stackoverflow.com/questions/27920690/linear-interpolation-using-dplyr/31845035#31845035}{SO Link}
}
\examples{
interpolate(1:5, c(10, NA, NA, NA, 100), "spline")

df <- dplyr::tibble(
  seq = 1:5,
  v1 = c(1, NA, 3, NA, 5),
  v2 = c(40, NA, 60, NA, 70),
  v3 = c(10, NA, NA, NA, 100))

df

df |>
   dplyr::mutate(
   dplyr::across(
   .cols = dplyr::starts_with("v"),
   .fns = ~ interpolate(seq, .x),
   .names = "{.col}_est"))

df |>
  dplyr::mutate(
   dplyr::across(
    dplyr::starts_with("v"),
     ~ interpolate(seq, .x)))

}
\seealso{
Other base: 
\code{\link{as_character}()},
\code{\link{colon}()},
\code{\link{invert_named}()}
}
\concept{base}
\keyword{maths}
