% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_named.R
\name{invert_named}
\alias{invert_named}
\title{Invert a named vector}
\usage{
invert_named(x)
}
\arguments{
\item{x}{A named vector}
}
\value{
A named vector with names and values inverted
}
\description{
Invert a named vector
}
\examples{
invert_named(x = c(name = "element"))

invert_named(x = c(element = "name"))

}
