% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nas.R
\name{iqr_na}
\alias{iqr_na}
\title{\code{IQR()} with \code{NA} removal}
\usage{
iqr_na(x, ...)
}
\arguments{
\item{x}{\verb{<numeric>} vector}

\item{...}{additional arguments to pass to \code{stats::IQR()}}
}
\description{
\code{IQR()} with \code{NA} removal
}
\examples{
iqr_na(c(1, 2, NA))

}
