% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_ops.R
\name{is_directory}
\alias{is_directory}
\title{Test if a path is a directory}
\usage{
is_directory(x)
}
\arguments{
\item{x}{\verb{<chr>} directory path to check}
}
\value{
named \verb{<lgl>} vector, where the names give the paths. If the given
object does not exist, \code{NA} is returned.
}
\description{
Test if a path is a directory
}
\examples{
is_directory("C:/")

}
