% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_ops.R
\name{is_readable}
\alias{is_readable}
\title{Test if a path is readable}
\usage{
is_readable(x)
}
\arguments{
\item{x}{\verb{<chr>} vector of one or more paths}
}
\value{
\verb{<lgl>} vector, with names corresponding to input path.
}
\description{
Test if a path is readable
}
\examples{
is_readable("D:/")

}
