% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{is_valid_npi2}
\alias{is_valid_npi2}
\title{Validate NPIs (v2)}
\usage{
is_valid_npi2(x)
}
\arguments{
\item{x}{\verb{<character>} Number of NPIs to generate}
}
\value{
\verb{<logical>} vector
}
\description{
Validate NPIs (v2)
}
\examples{
is_valid_npi2("1043477615")

is_valid_npi2("1234567891")

x <- random_npi_generator(100)

x[collapse::whichv(purrr::map_lgl(x, is_valid_npi2), TRUE)]

}
\keyword{checks}
