% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{list_pins}
\alias{list_pins}
\title{List pins from mount_board()}
\usage{
list_pins(...)
}
\arguments{
\item{...}{arguments to pass to mount_board()}
}
\value{
\verb{<chr>} vector of named pins
}
\description{
List pins from mount_board()
}
\keyword{internal}
