% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nas.R
\name{mad_na}
\alias{mad_na}
\title{\code{mad()} with \code{NA} removal}
\usage{
mad_na(x, ...)
}
\arguments{
\item{x}{\verb{<numeric>} vector}

\item{...}{additional arguments to pass to \code{stats::mad()}}
}
\description{
Compute the median absolute deviation, i.e., the (lo-/hi-) median of the
absolute deviations from the median, and (by default) adjust by a factor for
asymptotically normal consistency.
}
\examples{
mad_na(c(1, 2, NA))

}
