% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nas.R
\name{max_na}
\alias{max_na}
\title{\code{max()} with \code{NA} removal}
\usage{
max_na(...)
}
\arguments{
\item{...}{\verb{<numeric>} vector}
}
\description{
\code{max()} with \code{NA} removal
}
\examples{
max_na(c(1, 2, NA))

}
