% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{max_vlen}
\alias{max_vlen}
\title{Maximum Vector Length}
\usage{
max_vlen(x)
}
\arguments{
\item{x}{vector}
}
\value{
\verb{<int>} maximum length of vector
}
\description{
Maximum Vector Length
}
\keyword{internal}
