% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nas.R
\name{mean_na}
\alias{mean_na}
\title{\code{mean()} with \code{NA} removal}
\usage{
mean_na(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
\code{mean()} with \code{NA} removal
}
\examples{
mean_na(c(1, 2, NA))

}
