% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nas.R
\name{median_na}
\alias{median_na}
\title{\code{median()} with \code{NA} removal}
\usage{
median_na(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
\code{median()} with \code{NA} removal
}
\examples{
sd_na(c(1, 2, NA))

}
