% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nas.R
\name{min_na}
\alias{min_na}
\title{\code{min()} with \code{NA} removal}
\usage{
min_na(...)
}
\arguments{
\item{...}{\verb{<numeric>} vector}
}
\description{
\code{min()} with \code{NA} removal
}
\examples{
min_na(c(1, 2, NA))

}
