% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mock_forager}
\alias{mock_forager}
\title{Generate mock coding/billing data frame}
\usage{
mock_forager(rows = 10, nest = TRUE)
}
\arguments{
\item{rows}{\verb{<int>} number of rows to generate; default is 10}

\item{nest}{\verb{<lgl>} whether to nest the dates column; default is \code{TRUE}}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Generate mock coding/billing data frame
}
\examples{
mock_forager(rows = 5)

}
\seealso{
Other mock: 
\code{\link{mock_provider}()},
\code{\link{random_npi_generator}()}
}
\concept{mock}
