% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mock_provider}
\alias{mock_provider}
\title{Generate simple data frame for testing}
\usage{
mock_provider(years)
}
\arguments{
\item{years}{\verb{<int>} vector (\code{c(2020, 2021, 2023)}) or sequence of years
(\code{2010:2020})}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Generate simple data frame for testing
}
\examples{
mock_provider(2010:2020)

mock_provider(c(2020, 2021, 2023))

}
\seealso{
Other mock: 
\code{\link{mock_forager}()},
\code{\link{random_npi_generator}()}
}
\concept{mock}
