% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_clean.R
\name{months_regex}
\alias{months_regex}
\title{Regular expression for matching month names}
\usage{
months_regex()
}
\value{
\verb{<chr>} string  of a regex for matching month names
}
\description{
Regular expression for matching month names
}
\examples{
months_regex()

}
