% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastverse.R
\name{na}
\alias{na}
\title{Predicate to filter out NAs}
\usage{
na(x)
}
\arguments{
\item{x}{vector}
}
\value{
\verb{<logical>} vector
}
\description{
Predicate to filter out NAs
}
\examples{
na(c(NA, "AA"))

}
