% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_blank.R
\name{na_blank}
\alias{na_blank}
\title{Convert various character values to \code{NA}}
\usage{
na_blank(x)
}
\arguments{
\item{x}{\verb{<chr>} vector}
}
\value{
\verb{<chr>} vector with \code{NA} values
}
\description{
Convert various character values to \code{NA}
}
\examples{
na_blank(x = c(" ", "*", "--", "N/A", ""))

}
