% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{na_if}
\alias{na_if}
\title{Convert values to \code{NA}}
\usage{
na_if(x, y)
}
\arguments{
\item{x}{vector to modify}

\item{y}{values to convert to \code{NA}. Type must match that of \code{x}.}
}
\value{
\code{x} with values in \code{y} converted to \code{NA}
noRd
}
\description{
Convert values to \code{NA}
}
