% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_if_common.R
\name{na_if_common}
\alias{na_if_common}
\title{Convert various character strings to \code{NA}}
\usage{
na_if_common(x)
}
\arguments{
\item{x}{\verb{<chr>} vector to convert}
}
\value{
\verb{<chr>} vector with converted \code{NA} values
}
\description{
Convert various character strings to \code{NA}
}
\examples{
na_if_common(x = c(" ", "*", "--", "N/A", "", "A", "B"))

}
