% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{na_iqr}
\alias{na_iqr}
\title{\code{IQR()} with \code{NA} removal}
\usage{
na_iqr(x, ...)
}
\arguments{
\item{x}{\verb{<numeric>} vector}

\item{...}{additional arguments to pass to \code{stats::IQR()}}
}
\description{
\code{IQR()} with \code{NA} removal
}
\examples{
na_iqr(c(1, 2, NA))

}
