% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{na_mad}
\alias{na_mad}
\title{\code{mad()} with \code{NA} removal}
\usage{
na_mad(x, ...)
}
\arguments{
\item{x}{\verb{<numeric>} vector}

\item{...}{additional arguments to pass to \code{stats::mad()}}
}
\description{
Compute the median absolute deviation, i.e., the (lo-/hi-) median of the
absolute deviations from the median, and (by default) adjust by a factor for
asymptotically normal consistency.
}
\examples{
na_mad(c(1, 2, NA))

}
