% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{na_max}
\alias{na_max}
\title{\code{max()} with \code{NA} removal}
\usage{
na_max(...)
}
\arguments{
\item{...}{\verb{<numeric>} vector}
}
\description{
\code{max()} with \code{NA} removal
}
\examples{
na_max(c(1, 2, NA))

}
\keyword{internal}
