% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nas.R
\name{na_mean}
\alias{na_mean}
\title{\code{mean()} with \code{NA} removal}
\usage{
na_mean(x)
}
\arguments{
\item{x}{\verb{<numeric>} vector}
}
\description{
\code{mean()} with \code{NA} removal
}
\examples{
na_mean(c(1, 2, NA))

}
