% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{na_med}
\alias{na_med}
\title{\code{median()} with \code{NA} removal}
\usage{
na_med(x)
}
\arguments{
\item{x}{\verb{<numeric>} vector}
}
\description{
\code{median()} with \code{NA} removal
}
\examples{
na_med(c(1, 2, NA))

}
\keyword{internal}
