% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nas.R
\name{na_min}
\alias{na_min}
\title{\code{min()} with \code{NA} removal}
\usage{
na_min(...)
}
\arguments{
\item{...}{\verb{<numeric>} vector}
}
\description{
\code{min()} with \code{NA} removal
}
\examples{
na_min(c(1, 2, NA))

}
