% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nas.R
\name{na_range}
\alias{na_range}
\title{\code{range()} with \code{NA} removal}
\usage{
na_range(...)
}
\arguments{
\item{...}{\verb{<numeric>} vector}
}
\description{
\code{range()} with \code{NA} removal
}
\examples{
na_range(c(1, 2, NA))

}
