% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{na_sd}
\alias{na_sd}
\title{\code{sd()} with \code{NA} removal}
\usage{
na_sd(x)
}
\arguments{
\item{x}{\verb{<numeric>} vector}
}
\description{
\code{sd()} with \code{NA} removal
}
\examples{
na_sd(c(1, 2, NA))

}
