% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{na_sum}
\alias{na_sum}
\title{\code{sum()} with \code{NA} removal}
\usage{
na_sum(x)
}
\arguments{
\item{x}{\verb{<numeric>} vector}
}
\description{
\code{sum()} with \code{NA} removal
}
\examples{
na_sum(c(1, 2, NA))

}
\keyword{internal}
