% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{na_var}
\alias{na_var}
\title{\code{var()} with \code{NA} removal}
\usage{
na_var(x)
}
\arguments{
\item{x}{\verb{<numeric>} vector}
}
\description{
\code{var()} with \code{NA} removal
}
\examples{
na_var(c(1, 2, NA))

}
