% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chchchchanges.R
\name{new_value}
\alias{new_value}
\title{Calculate the new value from the old value and the percentage change between
the two values}
\usage{
new_value(old, percentage_change)
}
\arguments{
\item{old}{\verb{<dbl>} older value}

\item{percentage_change}{\verb{<dbl>} percentage change}
}
\value{
\verb{<dbl>} vector of new value
}
\description{
Calculate the new value from the old value and the percentage change between
the two values
}
\examples{
new_value(old = 942482, percentage_change = 0.007700943)

new_value(old = 1132783, percentage_change = -0.1679942)

# If the value of an object increased by 220\%, from the original value of
# $500,000, what is it worth now?

new_value(old = 500000, percentage_change = 2.2)

}
