% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{not_na}
\alias{not_na}
\title{Predicate to filter out NAs}
\usage{
not_na(x)
}
\arguments{
\item{x}{vector}
}
\value{
\verb{<lgl>} vector
}
\description{
Predicate to filter out NAs
}
\keyword{internal}
