% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{null}
\alias{null}
\title{Is \code{x} \code{NULL}?}
\usage{
null(x)
}
\arguments{
\item{x}{input}
}
\value{
\verb{<lgl>} \code{TRUE} if \code{x} is \code{NULL}, else \code{FALSE}
}
\description{
Is \code{x} \code{NULL}?
}
\examples{
null(NULL)

null("NULL")

}
\seealso{
Other base: 
\code{\link{as_character}()},
\code{\link{colon}()},
\code{\link{delist}()},
\code{\link{desplit}()},
\code{\link{interpolate}()},
\code{\link{invert_named}()},
\code{\link{not_null}()}
}
\concept{base}
\keyword{helpers}
