% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{null_if_empty}
\alias{null_if_empty}
\title{Return NULL if the input is empty, otherwise return the input}
\usage{
null_if_empty(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector or \code{NULL}
}
\description{
Return NULL if the input is empty, otherwise return the input
}
\examples{
null_if_empty(c())

null_if_empty(c(1))

}
