% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_npi_generator.R
\name{numeric_to_char}
\alias{numeric_to_char}
\title{Coerce \verb{<numeric>} vector to \verb{<character>}}
\usage{
numeric_to_char(x)
}
\arguments{
\item{x}{vector}
}
\description{
Coerce \verb{<numeric>} vector to \verb{<character>}
}
\keyword{internal}
