% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_ops.R
\name{pad_number}
\alias{pad_number}
\title{Pad numbers with zeroes}
\usage{
pad_number(x, digits = NULL)
}
\arguments{
\item{x}{\verb{<dbl>} vector of numbers}

\item{digits}{\verb{<int>} single integer; the number of zeroes to pad \code{x} with.
If left \code{NULL} (the default), will use the result of \code{max(nchar(x))}}
}
\value{
\verb{<chr>} vector of padded numbers
}
\description{
Pad numbers with zeroes
}
\examples{
pad_number(c(56, 2584, 010, 912, 12222))

pad_number(c(56, 2584, 010, 912, 00012222), digits = 10L)

}
