% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_ops.R
\name{parens}
\alias{parens}
\title{Wrapper for \code{\link[=paste0]{paste0()}} that adds parentheses}
\usage{
parens(x)
}
\arguments{
\item{x}{\verb{<chr>} string}
}
\description{
Wrapper for \code{\link[=paste0]{paste0()}} that adds parentheses
}
\keyword{internal}
