% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_change_diff.R
\name{pct_change}
\alias{pct_change}
\title{Calculate the percentage change between two values}
\usage{
pct_change(new, old)
}
\arguments{
\item{new}{\verb{<dbl>} newer value}

\item{old}{\verb{<dbl>} older value}
}
\value{
vector
}
\description{
Percentage change is obtained by dividing the change in value by the original
value. If the result is positive, then it is an increase. If the result is
negative, then it is a decrease.
}
\examples{
pct_change(new = 1, old = 100)

pct_change(new = 100, old = 1)

}
