% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_change_diff.R
\name{pct_difference}
\alias{pct_difference}
\title{Calculate the percent difference between two values}
\usage{
pct_difference(x, y)
}
\arguments{
\item{x, y}{\verb{<dbl>} values to determine the percent difference between.}
}
\value{
vector
}
\description{
Percent difference is obtained by dividing the absolute value of change by
the average of the values and then multiplying it with 100. Since the
absolute value is taken for the change (or difference) in values, the order
of the numbers does not matter.
}
\examples{
pct_difference(1, 100)

pct_difference(100, 1)

}
