% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chchchchanges.R
\name{percentage_calculator}
\alias{percentage_calculator}
\title{Percentage calculator}
\usage{
percentage_calculator(a, b)
}
\arguments{
\item{a, b}{\verb{<dbl>} values}
}
\value{
\verb{<dbl>} vector of percentage difference
}
\description{
If \emph{a} is some \emph{p\%} of \emph{b}, then:
\itemize{
\item \emph{p} = a/b
\item \emph{a} = p * b
\item \emph{b} = a/p
}
}
\details{
Since the absolute value is taken for the change (or difference) in values,
the order of the numbers does not matter.
}
\examples{
percentage_calculator(2500, 133)

percentage_calculator(133, 2500)

}
