% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percentage_change}
\alias{percentage_change}
\title{Calculate the percentage change between two values}
\usage{
percentage_change(new, old)
}
\arguments{
\item{new}{\verb{<dbl>} newer value}

\item{old}{\verb{<dbl>} older value}
}
\value{
\verb{<dbl>} vector of percentage change
}
\description{
Percentage change is obtained by dividing the change in value by the original
value.
}
\details{
If the result is positive, then it is an increase. If the result is negative,
then it is a decrease.
}
\examples{
percentage_change(new = 949740, old = 942482)

percentage_change(new = 942482, old = 1132783)

}
