% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chchchchanges.R
\name{percentage_difference}
\alias{percentage_difference}
\title{Calculate the percentage difference between two values}
\usage{
percentage_difference(x, y)
}
\arguments{
\item{x, y}{\verb{<dbl>} values to determine the percent difference between, order
does not matter}
}
\value{
\verb{<dbl>} vector of percentage difference
}
\description{
Percentage difference is obtained by dividing the absolute value of change by
the average of the values and then multiplying it with 100.
}
\details{
Since the absolute value is taken for the change (or difference) in values,
the order of the numbers does not matter.
}
\examples{
percentage_difference(949740, 942482)

percentage_difference(942482, 1132783)

}
