% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_list.R
\name{print_ls}
\alias{print_ls}
\title{Print a named list}
\usage{
print_ls(ls, prefix = "")
}
\arguments{
\item{ls}{\verb{<list>} to print}

\item{prefix}{\verb{<chr>} to prepend to each line}
}
\value{
\verb{<list>} invisibly
}
\description{
Print a named list
}
\examples{
print_ls(list(a = 1, b = 2, c = 3))

}
