% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_data.R
\name{provider_data}
\alias{provider_data}
\title{Generate Data for Testing}
\usage{
provider_data(year_seq)
}
\arguments{
\item{year_seq}{sequence of years, e.g. \code{2010:2020}}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Generate Data for Testing
}
\examples{
provider_data(2010:2020)

}
\keyword{internal}
