% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{random_npi_generator}
\alias{random_npi_generator}
\title{Generate random NPIs}
\usage{
random_npi_generator(n)
}
\arguments{
\item{n}{\verb{<integer>} Number of NPIs to generate}
}
\value{
\verb{<character>} vector of \code{n} NPIs
}
\description{
Generate random NPIs
}
\examples{
random_npi_generator(10)

x <- random_npi_generator(100)

x[collapse::whichv(purrr::map_lgl(x, is_valid_npi), TRUE)]

}
\seealso{
Other mock: 
\code{\link{forager_data}()},
\code{\link{provider_data}()}
}
\concept{mock}
