% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{random_string}
\alias{random_string}
\title{Generate random strings}
\usage{
random_string(n, ln, p = "[A-Z0-9]")
}
\arguments{
\item{n}{\verb{<int>} number of strings to generate}

\item{ln}{\verb{<int>} length of each string}

\item{p}{\verb{<chr>} pattern to sample from; default is \code{"[A-Z0-9]"}}
}
\value{
\verb{<chr>} vector of random strings
}
\description{
Generate random strings
}
\keyword{internal}
