% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nas.R
\name{range_na}
\alias{range_na}
\title{\code{range()} with \code{NA} removal}
\usage{
range_na(...)
}
\arguments{
\item{...}{\verb{<numeric>} vector}
}
\description{
\code{range()} with \code{NA} removal
}
\examples{
range_na(c(1, 2, NA))

}
