% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{remove_all_na}
\alias{remove_all_na}
\title{Remove columns and rows with all NAs}
\usage{
remove_all_na(x)
}
\arguments{
\item{x}{\verb{<data.frame>}}
}
\description{
Remove columns and rows with all NAs
}
