% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_quiet.R
\name{remove_quiet}
\alias{remove_quiet}
\title{Remove empty rows and columns}
\usage{
remove_quiet(df, ...)
}
\arguments{
\item{df}{data frame}

\item{...}{additional arguments to pass to \code{janitor::remove_empty()}}
}
\description{
Remove empty rows and columns
}
\examples{
x <- dplyr::tibble(
  x = c(1, 2, NA),
  y = c(NA, NA, NA))

x

remove_quiet(x)

}
