% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{remove_quotes}
\alias{remove_quotes}
\title{Remove single or double quotes from a character string}
\usage{
remove_quotes(x)
}
\arguments{
\item{x}{\verb{<chr>} vector to convert}
}
\value{
\verb{<chr>} vector with quotes removed
}
\description{
Remove single or double quotes from a character string
}
\keyword{internal}
