% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{replace_fixed}
\alias{replace_fixed}
\title{Replace All Fixed Strings}
\usage{
replace_fixed(x, p, r, v = FALSE)
}
\arguments{
\item{x}{\verb{<chr>} vector or string to search in}

\item{p}{\verb{<chr>} vector of search patterns}

\item{r}{\verb{<chr>} vector of replacements for matched patterns}

\item{v}{\verb{<lgl>} default is FALSE; should each occurrence of a pattern in
every string be replaced by a corresponding replacement string?}
}
\value{
\verb{<chr>} vector with all occurrences of each pattern replaced
}
\description{
Replace All Fixed Strings
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
replace_fixed(
  x,
  c(":", "\%", "@", "$"),
  c("_", "", "", ""))
\dontshow{\}) # examplesIf}
}
\keyword{internal}
