% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{roundup}
\alias{roundup}
\title{Round up to nearest decimal place}
\usage{
roundup(x, d = 2)
}
\arguments{
\item{x}{\verb{<num>} vector}

\item{d}{\verb{<int>} decimal places to round to; default is \code{2}}
}
\value{
\verb{<num>} vector rounded up to the nearest decimal place
}
\description{
Round up to nearest decimal place
}
\keyword{internal}
