% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{search_for}
\alias{search_for}
\title{Search a data frame column by string}
\usage{
search_for(data, column, search, ignore = TRUE, ...)
}
\arguments{
\item{data}{\verb{<data.frame>} or \verb{<tibble>}}

\item{column}{\verb{<chr>} string, column to search}

\item{search}{\verb{<chr>} string to search for, can be a regex}

\item{ignore}{\verb{<lgl>} ignore case, default is \code{TRUE}}

\item{...}{additional arguments}
}
\value{
\verb{<data.frame>} or \verb{<tibble>}
}
\description{
Search a data frame column by string
}
\examples{
(x <- dplyr::tibble(y = 1:10, z = letters[1:10]))

search_for(data = x, column = "z", search = "[a|j]")

}
\seealso{
Other tidyverse: 
\code{\link{columns_to_character}()},
\code{\link{gchop}()},
\code{\link{glue_chr}()},
\code{\link{glue_data_chr}()},
\code{\link{named_group_split}()}
}
\concept{tidyverse}
\keyword{helpers}
