% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs.R
\name{search_in}
\alias{search_in}
\title{Search in data frame column if search term is not \code{NULL}}
\usage{
search_in(data, column, what)
}
\arguments{
\item{data}{\verb{<data.frame>} or \verb{<tibble>}}

\item{column}{column to search; of the form \code{data[["column"]]} or
\code{data$column}}

\item{what}{string or vector to search for in \code{column}}
}
\value{
If \code{what} is not \code{NULL}, \verb{<data.frame>} or \verb{<tibble>} with rows
containing \code{what}
}
\description{
Search in data frame column if search term is not \code{NULL}
}
\examples{
x <- dplyr::tibble(
     y = 1:10,
     z = letters[1:10])

search_in(data = x,
          column = x[["z"]],
          what = c("a", "j"))

search_in(data = x,
          column = x[["z"]],
          what = NULL)

}
\seealso{
Other vctrs: 
\code{\link{empty}()},
\code{\link{gchop}()},
\code{\link{null_if_empty}()},
\code{\link{search_in_impl}()}
}
\concept{vctrs}
\keyword{programming}
