% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_in_if.R
\name{search_in_if}
\alias{search_in_if}
\title{Search in data frame column if search term is not \code{NULL}}
\usage{
search_in_if(df, dfcol, search)
}
\arguments{
\item{df}{A \verb{<data.frame>} or \verb{<tibble>}}

\item{dfcol}{A \verb{<character>} or \verb{<symbol>} specifying the column to search in}

\item{search}{A \verb{<character>} or \verb{<symbol>} specifying the search term}
}
\value{
A \verb{<data.frame>} or \verb{<tibble>}
}
\description{
Search in data frame column if search term is not \code{NULL}
}
\examples{
x <- dplyr::tibble(
     y = 1:10,
     z = letters[1:10])

search_in_if(df = x,
             dfcol = x$z,
             search = c("a", "j"))

search_in_if(df = x,
             dfcol = x$z,
             search = NULL)

}
