% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs.R
\name{search_in_impl}
\alias{search_in_impl}
\title{Search in data frame}
\usage{
search_in_impl(data, column, what)
}
\arguments{
\item{data}{A \verb{<data.frame>} or \verb{<tibble>}}

\item{column}{column to search; of the form \code{data[["column"]]} or
\code{data$column}}

\item{what}{string or vector to search for in \code{column}}
}
\value{
\verb{<data.frame>} or \verb{<tibble>} with rows containing \code{what}
}
\description{
Search in data frame
}
\examples{
x <- dplyr::tibble(y = 1:10,
                   z = letters[1:10])

search_in_impl(data = x,
               column = x[["z"]],
               what = c("a", "j"))

}
\seealso{
Other vctrs: 
\code{\link{empty}()},
\code{\link{gchop}()},
\code{\link{null_if_empty}()},
\code{\link{search_in}()}
}
\concept{vctrs}
\keyword{programming}
