% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{sf_at}
\alias{sf_at}
\title{Subset Vector at One Index Point}
\usage{
sf_at(x, idx = 1)
}
\arguments{
\item{x}{\verb{<chr>} vector}

\item{idx}{\verb{<int>} index}
}
\value{
\verb{<chr>} vector
}
\description{
Subset Vector at One Index Point
}
\keyword{internal}
