% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringfish.R
\name{sf_c}
\alias{sf_c}
\title{Concatenate Vectors}
\usage{
sf_c(...)
}
\arguments{
\item{...}{Any number of vectors, coerced to \verb{<character>} vector, if necessary}
}
\value{
concatenated \verb{<character>} vector
}
\description{
Concatenate Vectors
}
\examples{
sf_c(LETTERS, "A")

}
