% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{sf_chars}
\alias{sf_chars}
\title{Count number of characters in character vector}
\usage{
sf_chars(x)
}
\arguments{
\item{x}{\verb{<chr>} vector}
}
\value{
\verb{<int>} vector
}
\description{
Count number of characters in character vector
}
\keyword{internal}
