% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{sf_conv}
\alias{sf_conv}
\title{Convert string to stringfish vector}
\usage{
sf_conv(x)
}
\arguments{
\item{x}{\verb{<chr>} vector}
}
\value{
\verb{<chr>} vector
}
\description{
Convert string to stringfish vector
}
\keyword{internal}
