% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{sf_extract}
\alias{sf_extract}
\title{Extract by Regex}
\usage{
sf_extract(s, p)
}
\arguments{
\item{s}{\verb{<chr>} vector}

\item{p}{\verb{<chr>} regex pattern}
}
\value{
\verb{<chr>} vector with pattern extracted
}
\description{
Extract by Regex
}
\keyword{internal}
