% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringfish.R
\name{sf_nchar}
\alias{sf_nchar}
\title{Count number of characters in character vector}
\usage{
sf_nchar(x)
}
\arguments{
\item{x}{\verb{<character>} vector}
}
\value{
\verb{<character>} stringfish vector
}
\description{
Count number of characters in character vector
}
\examples{
sf_nchar(random_npi_generator(10))

}
