% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringfish.R
\name{sf_ndetect}
\alias{sf_ndetect}
\title{Detect Opposite by Regex}
\usage{
sf_ndetect(s, p)
}
\arguments{
\item{s}{\verb{<character>} vector}

\item{p}{\verb{<character>} regex pattern}
}
\value{
\verb{<logical>} vector
}
\description{
Detect Opposite by Regex
}
\examples{
sf_ndetect(random_npi_generator(10), "[A-Z]{1}")

}
