% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringfish.R
\name{sf_nextract}
\alias{sf_nextract}
\title{Extract Opposite by Regex}
\usage{
sf_nextract(s, p)
}
\arguments{
\item{s}{\verb{<character>} vector}

\item{p}{\verb{<character>} regex pattern}
}
\value{
\verb{<character>} vector
}
\description{
Extract Opposite by Regex
}
\examples{
sf_nextract(sf_c(LETTERS, 0:9), "[A-Z]")

sf_nextract(random_npi_generator(10), "[A-Z]")

}
