% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{sf_remove}
\alias{sf_remove}
\title{Remove by Regex}
\usage{
sf_remove(s, p, fix = FALSE)
}
\arguments{
\item{s}{\verb{<chr>} vector}

\item{p}{\verb{<chr>} regex pattern}

\item{fix}{\verb{<lgl>} fixed or regex; default is \code{FALSE}}
}
\value{
\verb{<chr>} vector with pattern removed
}
\description{
Remove by Regex
}
\keyword{internal}
