% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{sf_replace}
\alias{sf_replace}
\title{Replace by Regex}
\usage{
sf_replace(s, p, r, fix = FALSE)
}
\arguments{
\item{s}{\verb{<chr>} vector}

\item{p}{\verb{<chr>} regex pattern}

\item{r}{\verb{<chr>} replacement}

\item{fix}{\verb{<lgl>} fixed or regex; default is \code{FALSE}}
}
\value{
\verb{<chr>} vector with pattern replaced
}
\description{
Replace by Regex
}
\keyword{internal}
