% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringfish.R
\name{sf_smush}
\alias{sf_smush}
\title{Collapse Vector}
\usage{
sf_smush(x, sep = "")
}
\arguments{
\item{x}{\verb{<character>} vector}

\item{sep}{\verb{<character>} separator; default is \code{""}}
}
\value{
collapsed \verb{<character>} vector
}
\description{
Collapse Vector
}
\examples{
sf_smush(LETTERS, "|")

}
