% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringfish.R
\name{sf_strsplit}
\alias{sf_strsplit}
\title{Split String by Delimiter}
\usage{
sf_strsplit(x, s, fixed = TRUE)
}
\arguments{
\item{x}{\verb{<character>} vector}

\item{s}{\verb{<character>} delimiter to split by}

\item{fixed}{\verb{<logical>} fixed or regex; default is \code{TRUE}}
}
\value{
\verb{<list>} of character vectors
}
\description{
Split String by Delimiter
}
\examples{
sf_strsplit("A|B|C", "|")

}
