% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringfish.R
\name{sf_sub}
\alias{sf_sub}
\title{Subset Vector by Range}
\usage{
sf_sub(x, start = 1, stop)
}
\arguments{
\item{x}{\verb{<character>} vector}

\item{start}{\verb{<integer>} index start; default is \code{1}}

\item{stop}{\verb{<integer>} index end}
}
\value{
\verb{<character>} vector
}
\description{
Subset Vector by Range
}
\examples{
sf_sub(random_npi_generator(10), 1, 2)

}
