% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{show_missing}
\alias{show_missing}
\title{Show missing values}
\usage{
show_missing(df, group_vars, summary_vars = dplyr::everything())
}
\arguments{
\item{df}{A \verb{<data.frame>} or \verb{<tibble>}}

\item{group_vars}{A \verb{<character>} or \verb{<symbol>} vector of the variables to
group by}

\item{summary_vars}{A \verb{<character>} or \verb{<symbol>} vector of the variables to
summarize; default is \code{\link[dplyr:reexports]{dplyr::everything()}}}
}
\description{
Show missing values
}
\examples{
dplyr::tibble(x = 1:10,
              y = 1:10,
              z = c(letters[1:10])) |>
              show_missing(z)

}
