% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_ops.R
\name{single_line_string}
\alias{single_line_string}
\title{Format multiple line character vector to single line}
\usage{
single_line_string(x)
}
\arguments{
\item{x}{\verb{<chr>} character vector with line breaks (\verb{\\n})}
}
\value{
\verb{<chr>} single line character vector
}
\description{
Format multiple line character vector to single line
}
\examples{
single_line_string(
"(Jan(?:uary)?|Feb(?:ruary)?|Mar(?:ch)?|
Apr(?:il)?|May|Jun(?:e)?|Jul(?:y)?|")

}
