% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_gt.R
\name{sorted_bars}
\alias{sorted_bars}
\title{Sorted Bar Chart}
\usage{
sorted_bars(df, var)
}
\arguments{
\item{df}{data frame}

\item{var}{column to plot}
}
\value{
A \link[ggplot2:ggplot2-package]{ggplot2} object
}
\description{
Sorted Bar Chart
}
\examples{
ggplot2::diamonds |>
  sorted_bars(cut) +
  gg_theme()

}
