% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ins.R
\name{srchcol}
\alias{srchcol}
\title{Search a data frame column by string}
\usage{
srchcol(df, col, search, ignore = TRUE, ...)
}
\arguments{
\item{df}{\verb{<data.frame>} or \verb{<tibble>}}

\item{col}{column name as string or unquoted}

\item{search}{string to search for, can be a regular expression, e.g. \code{'^[A-Z]'}}

\item{ignore}{ignore case, default is \code{TRUE}}

\item{...}{additional arguments}
}
\value{
\verb{<data.frame>} or \verb{<tibble>}
}
\description{
Search a data frame column by string
}
\examples{
x <- dplyr::tibble(y = 1:10, z = letters[1:10])

srchcol(df = x, col = "z", search = "[a|j]")

}
