% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{strsort}
\alias{strsort}
\title{Parallel Sort}
\usage{
strsort(x)
}
\arguments{
\item{x}{\verb{<chr>} vector. If other, will default to \code{\link[base:sort]{base::sort()}}}
}
\value{
\verb{<chr>} vector
}
\description{
Parallel Sort
}
\keyword{internal}
