% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nas.R
\name{sum_na}
\alias{sum_na}
\title{\code{sum()} with \code{NA} removal}
\usage{
sum_na(x)
}
\arguments{
\item{x}{\verb{<numeric>} vector}
}
\description{
\code{sum()} with \code{NA} removal
}
\examples{
sum_na(c(1, 2, NA))

}
