% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{summary_stats}
\alias{summary_stats}
\title{Summary statistics}
\usage{
summary_stats(
  df,
  condition = NULL,
  group_vars = NULL,
  summary_vars = NULL,
  arr = NULL,
  digits = 3
)
}
\arguments{
\item{df}{data frame}

\item{condition}{filter condition, i.e. \code{patient == "new"}}

\item{group_vars}{variables to group by, i.e. \code{c(specialty, state, hcpcs, cost)}}

\item{summary_vars}{variables to summarise, i.e. \code{c(min, max, mode, range)}}

\item{arr}{column to arrange data by, i.e. \code{cost}}

\item{digits}{Number of digits to round to, default is 3}
}
\value{
A \verb{<tibble>} with the summarized data
}
\description{
Summary statistics
}
\examples{
x <- dplyr::tibble(
   provider = sample(c("A", "B", "C"), size = 200, replace = TRUE),
   city = sample(c("ATL", "NYC"), size = 200, replace = TRUE),
   charges = sample(1000:2000, size = 200),
   payment = sample(1000:2000, size = 200))

summary_stats(
   x,
   condition    = city == "ATL",
   group_vars   = provider,
   summary_vars = c(charges, payment),
   arr          = provider)
}
