% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{true}
\alias{true}
\title{Is \code{x} \code{TRUE}?}
\usage{
true(x)
}
\arguments{
\item{x}{input}
}
\value{
\verb{<lgl>} \code{TRUE} if \code{x} is \code{TRUE}, else \code{FALSE}
}
\description{
Is \code{x} \code{TRUE}?
}
\keyword{internal}
