% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{ttimestamp}
\alias{ttimestamp}
\title{Timestamp}
\usage{
ttimestamp(time = Sys.time())
}
\arguments{
\item{time}{Sys.time() or a date-time object}
}
\value{
A list()
}
\description{
Timestamp
}
\examples{
timestamp()

date()

ttimestamp()

}
