% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastverse.R
\name{uniq}
\alias{uniq}
\title{Unique Values of Vector}
\usage{
uniq(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector
}
\description{
Unique Values of Vector
}
\examples{
random_npi_generator(10) |> uniq()

}
