% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{uniq_rmna}
\alias{uniq_rmna}
\title{Remove duplicates and \code{NA} values}
\usage{
uniq_rmna(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector with duplicates and \code{NA} values removed
}
\description{
Remove duplicates and \code{NA} values
}
\examples{
c("4", "100", "100", NA)

uniq_rmna(c("4", "100", "100", NA))

}
