% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{uniq_vlen}
\alias{uniq_vlen}
\title{Unique Lengths of Vector}
\usage{
uniq_vlen(x)
}
\arguments{
\item{x}{vector}
}
\value{
\verb{<int>} vector
}
\description{
Unique Lengths of Vector
}
\keyword{internal}
