% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_development.R
\name{update_personal_packages}
\alias{update_personal_packages}
\title{Install/update personal packages}
\usage{
update_personal_packages()
}
\description{
Install/update personal packages
}
\keyword{internal}
