% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nas.R
\name{var_na}
\alias{var_na}
\title{\code{var()} with \code{NA} removal}
\usage{
var_na(x)
}
\arguments{
\item{x}{\verb{<numeric>} vector}
}
\description{
\code{var()} with \code{NA} removal
}
\examples{
var_na(c(1, 2, NA))

}
