% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastverse.R
\name{vlen}
\alias{vlen}
\title{Lengths of Vector}
\usage{
vlen(x)
}
\arguments{
\item{x}{vector}
}
\value{
\verb{<integer>} vector
}
\description{
Lengths of Vector
}
\examples{
random_npi_generator(10) |> vlen()

}
