% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_calculations.R
\name{years_df}
\alias{years_df}
\title{Calculate the number of years between two dates}
\usage{
years_df(df, from, to)
}
\arguments{
\item{df}{A \verb{<data.frame>} or \verb{<tibble>}}

\item{from}{\verb{<date>} Start date column}

\item{to}{\verb{<date>} End date column}
}
\value{
\verb{<data.frame>} or \verb{<tibble>}
}
\description{
Calculate the number of years between two dates
}
\examples{
dplyr::tibble(
  from = as.Date("2020-01-01"),
  to = as.Date("2020-01-01") + 2057) |>
  years_df(from, to)

}
