% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years_floor.R
\name{years_floor}
\alias{years_floor}
\title{Calculate the number of years between two dates rounded down}
\usage{
years_floor(from, to)
}
\arguments{
\item{from}{\verb{<date>} Start date}

\item{to}{\verb{<date>} End date}
}
\value{
\verb{<dbl>} vector of the number of years between the two dates
}
\description{
Calculate the number of years between two dates rounded down
}
\examples{
years_floor(
   from = as.Date("2020-01-01"),
   to   = as.Date("2020-01-01") + 2057)

}
