% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{years_vec}
\alias{years_vec}
\title{Calculate the number of years between two dates}
\usage{
years_vec(from, to)
}
\arguments{
\item{from}{\verb{<date>} Start date}

\item{to}{\verb{<date>} End date}
}
\value{
\verb{<dbl>} vector of the number of years between the two dates
}
\description{
Calculate the number of years between two dates
}
\examples{
years_vec(
  from = as.Date("2020-01-01"),
  to = as.Date("2020-01-01") + 2057)

}
