% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{as_character}
\alias{as_character}
\title{Coerce vector to \verb{<character>}}
\usage{
as_character(x)
}
\arguments{
\item{x}{vector}
}
\value{
\verb{<character>} vector
}
\description{
Coerce vector to \verb{<character>}
}
\examples{
as_character(NULL)

as_character(123)

}
\seealso{
Other base: 
\code{\link{colon}()},
\code{\link{delist}()},
\code{\link{desplit}()},
\code{\link{interpolate}()},
\code{\link{invert_named}()},
\code{\link{not_null}()},
\code{\link{null}()}
}
\concept{base}
\keyword{helpers}
