% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chchchchanges.R
\name{change_lagged}
\alias{change_lagged}
\title{Calculate lagged values by column}
\usage{
change_lagged(df, col, by = NULL)
}
\arguments{
\item{df}{data frame}

\item{col}{column of numeric values to calculate lag}

\item{by}{column to calculate lag by}
}
\description{
Calculate lagged values by column
}
\examples{
provider_data(2020:2025) |>
dplyr::group_by(group) |>
change_lagged(net_payment, year)

}
