% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{colon}
\alias{colon}
\title{Create Integer Sequence Beginning at 1}
\usage{
colon(n)
}
\arguments{
\item{n}{\verb{<integer>} Ending number of sequence}
}
\value{
\verb{<integer>} vector of numbers from 1 to \code{n}
}
\description{
Create Integer Sequence Beginning at 1
}
\examples{
colon(50)

colon(-20)

colon(0 + 150 - 145)

colon(20.9)

colon(20.1)

}
\seealso{
Other base: 
\code{\link{as_character}()},
\code{\link{interpolate}()},
\code{\link{invert_named}()}
}
\concept{base}
\keyword{helpers}
