% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine multiple columns into one}
\usage{
combine(df, name = combined, columns, sep = "-")
}
\arguments{
\item{df}{\verb{<data.frame>} or \verb{<tibble>}}

\item{name}{new column name, unquoted, default is \code{combined}}

\item{columns}{\verb{<chr>} vector of columns to combine}

\item{sep}{separator between combined row data, default is \code{"-"}}
}
\value{
A \verb{<data.frame>} or \verb{<tibble>} with combined columns
}
\description{
Combine multiple columns into one
}
\examples{
fuimus:::forager_data()[-5] |>
  combine(name = id_payer,
  columns = c('claim_id', 'payer'))

}
