% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{combine}
\alias{combine}
\title{Combine multiple columns into one}
\usage{
combine(data, name = combined, columns, sep = "-")
}
\arguments{
\item{data}{\verb{<data.frame>} or \verb{<tibble>}}

\item{name}{new column name, unquoted, default is \code{combined}}

\item{columns}{\verb{<chr>} vector of columns to combine}

\item{sep}{separator between combined row data, default is \code{"-"}}
}
\value{
A \verb{<tibble>} with combined columns
}
\description{
Combine multiple columns into one
}
\examples{
x <- forager_data()[-5]

x[1, 2] <- ""

x

combine(
  data = x,
  name = payer,
  columns = c('claim_id', 'payer'))

}
