% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{count_prop}
\alias{count_prop}
\title{Count and calculate proportion of each group}
\usage{
count_prop(df, var, sort = FALSE, na.rm = FALSE)
}
\arguments{
\item{df}{A \verb{<data.frame>} or \verb{<tibble>}}

\item{var}{A \verb{<character>} or \verb{<symbol>} specifying the column to count}

\item{sort}{A \verb{<logical>} indicating whether to sort the output by frequency,
default is \code{FALSE}}

\item{na.rm}{A \verb{<logical>} indicating whether to remove missing values from
the count, default is \code{FALSE}}
}
\description{
Count and calculate proportion of each group
}
\examples{
forager_data() |>
  count_prop(payer, sort = TRUE)

}
