% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{delist}
\alias{delist}
\title{Unlist and unname}
\usage{
delist(x)
}
\arguments{
\item{x}{Named or unnamed \verb{<list>}}
}
\value{
Unnamed \verb{<chr>} vector
}
\description{
Unlist and unname
}
\keyword{internal}
