% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_misc.R
\name{density}
\alias{density}
\title{Density plot}
\usage{
density(df, var, color, facets, binwidth = 0.1)
}
\arguments{
\item{df}{df}

\item{var}{var}

\item{color}{color var}

\item{facets}{facets var}

\item{binwidth}{binwidth}
}
\value{
ggplot2 density plot
}
\description{
Density plot
}
\examples{
ggplot2::diamonds |> density(carat)

ggplot2::diamonds |> density(carat, cut)

ggplot2::diamonds |> density(carat, cut, clarity)

}
