% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_unique}
\alias{describe_unique}
\title{Describe unique values in a data frame}
\usage{
describe_unique(
  df,
  ...,
  .rename_first = "value",
  .set_names = NULL,
  .names_to = "variable"
)
}
\arguments{
\item{df}{data frame}

\item{...}{unquoted data frame columns}

\item{.rename_first}{default is \code{"value"}}

\item{.set_names}{vector of names}

\item{.names_to}{default is \code{"variable"}}
}
\value{
a data frame with unique values
}
\description{
Describe unique values in a data frame
}
\examples{

describe_unique(fuimus:::forager_data(), claim_id, payer)

# describe_unique(fuimus:::forager_data(200), names(df)[2:3])

}
