% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{desplit}
\alias{desplit}
\title{Delist, Unname and Split a String}
\usage{
desplit(x)
}
\arguments{
\item{x}{\verb{<character>} string or named \verb{<list>}}
}
\value{
Unnamed \verb{<list>} of split \verb{<character>} vectors
}
\description{
Delist, Unname and Split a String
}
\examples{
# unnamed vector
desplit("XYZ")

# named vector
desplit(c(x = "XYZ"))

# unnamed list with one element
desplit(list("XYZ"))

# unnamed list with multiple elements
desplit(list("YYY", "ZZZ"))

# named list with one element
desplit(list(x = "XYZ"))

# named list with multiple elements
desplit(list(x = "YYY", xx = "ZZZ"))

}
\seealso{
Other base: 
\code{\link{as_character}()},
\code{\link{colon}()},
\code{\link{delist}()},
\code{\link{interpolate}()},
\code{\link{invert_named}()},
\code{\link{not_null}()},
\code{\link{null}()}
}
\concept{base}
\keyword{helpers}
