% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_2_chr.R
\name{df_2_chr}
\alias{df_2_chr}
\title{Coerce numeric columns to character}
\usage{
df_2_chr(df)
}
\arguments{
\item{df}{\verb{<data.frame>} or \verb{<tibble>}}
}
\value{
A \verb{<data.frame>} or \verb{<tibble>} with numeric columns coerced to character
}
\description{
Coerce numeric columns to character
}
\examples{
df_2_chr(dplyr::tibble(x = 20:35))

}
