% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{display_long}
\alias{display_long}
\title{Pivot data frame to long format for easy printing}
\usage{
display_long(df, cols = dplyr::everything())
}
\arguments{
\item{df}{\verb{<data.frame>} or \verb{<tibble>} to pivot long}

\item{cols}{\verb{<character>} vector of columns to pivot long, default is \code{\link[dplyr:reexports]{dplyr::everything()}}}
}
\value{
a \verb{<data.frame>} or \verb{<tibble>} in long format
}
\description{
Pivot data frame to long format for easy printing
}
\examples{
dplyr::tibble(
   int = 1:10,
   chr = letters[1:10],
   date = rep(Sys.Date()), nrow(int)) |>
display_long()

}
