% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs.R
\name{empty}
\alias{empty}
\title{Is Vector Empty?}
\usage{
empty(x)
}
\arguments{
\item{x}{vector}
}
\value{
\verb{<logical>}
}
\description{
Is Vector Empty?
}
\examples{
empty(character(0))

empty(c())

empty(NULL)

empty(NA)

empty(list())

empty(list(character(0)))

empty(list(x = character(0)))

}
\seealso{
Other vctrs: 
\code{\link{gchop}()},
\code{\link{null_if_empty}()},
\code{\link{search_in}()},
\code{\link{search_in_impl}()}
}
\concept{vctrs}
