% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_data.R
\name{forager_data}
\alias{forager_data}
\title{Generate mock coding/billing data frame}
\usage{
forager_data(rows = 10)
}
\arguments{
\item{rows}{number of rows to generate; default is 10}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Generate mock coding/billing data frame
}
\examples{
forager_data(rows = 5)

}
\keyword{internal}
