% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes.R
\name{geomean}
\alias{geomean}
\title{Calculate Geometric Mean}
\usage{
geomean(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Average \code{\link[=rate_of_return]{rate_of_return()}}
}
\examples{
x <- fuimus:::provider_data(2020:2025)

x |>
dplyr::group_by(group) |>
rate_of_return(net_payment)

x |>
dplyr::group_by(group) |>
rate_of_return(net_payment) |>
dplyr::summarise(gmean = geomean(net_payment_ror),
.by = group)

}
