% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastverse.R
\name{getelem}
\alias{getelem}
\title{Get named element from list}
\usage{
getelem(l, nm)
}
\arguments{
\item{l}{named \verb{<list>}}

\item{nm}{\verb{<character>} element name; can be a regex pattern}
}
\value{
named \verb{<list>} element
}
\description{
Get named element from list
}
\examples{
list(x1 = NA, x2 = "AA") |> getelem("x2")

list(x1 = NA, x2 = "AA") |> getelem("2$")

}
