% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{glue_chr}
\alias{glue_chr}
\title{Coerce glue() output to character}
\usage{
glue_chr(...)
}
\arguments{
\item{...}{dots to pass to glue function}
}
\value{
\verb{<chr>} vector
}
\description{
Coerce glue() output to character
}
\seealso{
Other tidyverse: 
\code{\link{columns_to_character}()},
\code{\link{gchop}()},
\code{\link{glue_data_chr}()},
\code{\link{named_group_split}()},
\code{\link{search_for}()}
}
\concept{tidyverse}
\keyword{helpers}
