% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{glue_chr}
\alias{glue_chr}
\title{Wrapper for \code{as.character(glue::glue(x))}}
\usage{
glue_chr(...)
}
\arguments{
\item{...}{dots to pass to glue function}
}
\value{
\verb{<chr>} vector
}
\description{
Wrapper for \code{as.character(glue::glue(x))}
}
\keyword{internal}
