% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_ops.R
\name{glue_chr}
\alias{glue_chr}
\title{Apply as.character() to glue()}
\usage{
glue_chr(...)
}
\arguments{
\item{...}{dots to pass to glue function}
}
\value{
\verb{<chr>} vector
}
\description{
Apply as.character() to glue()
}
