% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_ops.R
\name{glue_data_chr}
\alias{glue_data_chr}
\title{Wrapper for \code{as.character(glue_data(x))}}
\usage{
glue_data_chr(.x, ...)
}
\arguments{
\item{.x}{\verb{<vec>} vector to pass to glue_data()}

\item{...}{dots to pass to glue function}
}
\value{
\verb{<chr>} vector
}
\description{
Wrapper for \code{as.character(glue_data(x))}
}
