% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Infix operator for \verb{not in} statements}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{vector}

\item{table}{vector}
}
\value{
logical vector
}
\description{
Infix operator for \verb{not in} statements
}
