% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{\%nn\%}
\alias{\%nn\%}
\title{Infix operator for \code{if (!is.null(x)) y else x} statements}
\usage{
x \%nn\% y
}
\arguments{
\item{x, y}{vectors}
}
\value{
\code{y} if \code{x} is not \code{NULL}, else \code{x}
}
\description{
Infix operator for \code{if (!is.null(x)) y else x} statements
}
\examples{
NULL \%nn\% 123456L

"abc" \%nn\% 123456L

}
