% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_misc.R
\name{histogram}
\alias{histogram}
\title{Histogram}
\usage{
histogram(df, var, binwidth)
}
\arguments{
\item{df}{data frame}

\item{var}{var}

\item{binwidth}{\verb{<numeric>} binwidth}
}
\value{
ggplot2 histogram
}
\description{
Histogram
}
\examples{
ggplot2::diamonds |>
  histogram(carat, 0.1)

}
