% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{iif_else}
\alias{iif_else}
\title{ifelse wrapper using kit::iif}
\usage{
iif_else(x, yes, no)
}
\arguments{
\item{x}{\verb{<lgl>} vector}

\item{yes, no}{Values to return depending on TRUE/FALSE element of \code{x}. Must
be same type and be either length 1 or same length of \code{x}.}
}
\value{
\verb{<chr>} vector
}
\description{
ifelse wrapper using kit::iif
}
\keyword{internal}
