% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{invert_named}
\alias{invert_named}
\title{Invert a named vector}
\usage{
invert_named(x)
}
\arguments{
\item{x}{A named vector}
}
\value{
A named vector with names and values inverted
}
\description{
Invert a named vector
}
\examples{
invert_named(x = c(name = "element"))

invert_named(x = c(element = "name"))

}
\seealso{
Other base: 
\code{\link{as_character}()},
\code{\link{colon}()},
\code{\link{delist}()},
\code{\link{desplit}()},
\code{\link{interpolate}()},
\code{\link{not_null}()},
\code{\link{null}()}
}
\concept{base}
\keyword{helpers}
