% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastverse.R
\name{max_vlen}
\alias{max_vlen}
\title{Maximum Vector Length}
\usage{
max_vlen(x)
}
\arguments{
\item{x}{vector}
}
\value{
\verb{<integer>} vector
}
\description{
Maximum Vector Length
}
\examples{
random_npi_generator(10) |> max_vlen()

}
