% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_group_split.R
\name{named_group_split}
\alias{named_group_split}
\title{Split a \code{tibble} by Groups with Named List Output}
\usage{
named_group_split(df, ...)
}
\arguments{
\item{df}{\verb{<tibble>} or \verb{<data.frame>} to split}

\item{...}{One or more unquoted variables by which to group and then split
\code{df}. Variables can be separated by commas.}
}
\value{
named \verb{<list>} of tibbles, the names of which are derived from the
unique combinations of grouping variables, separated by "_".
}
\description{
This function takes a table, groups it by one or more variables, and then
splits the grouped data into a list. The resulting list has names derived
from the unique combinations of the grouping variables.
}
\examples{
x <- dplyr::tibble(
  zip = c("Data_Weekly.zip",
        "Data_Weekly.zip",
        "Data_April.zip",
        "Deactivated.zip"),
  file = c(
    "npidata.csv",
    "npidata2.csv",
    "endpoint.csv",
    "Deactivated.xlsx"))

x

named_group_split(x, zip)

}
