% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{null}
\alias{null}
\title{Is \code{x} \code{NULL}?}
\usage{
null(x)
}
\arguments{
\item{x}{input}
}
\value{
\verb{<lgl>} \code{TRUE} if \code{x} is \code{NULL}, else \code{FALSE}
}
\description{
Is \code{x} \code{NULL}?
}
\keyword{internal}
