% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{provider_data}
\alias{provider_data}
\title{Generate Data for Testing}
\usage{
provider_data(year_seq)
}
\arguments{
\item{year_seq}{sequence of years, e.g. \code{2010:2020}}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Generate Data for Testing
}
\examples{
provider_data(2010:2020)

}
\seealso{
Other mock: 
\code{\link{forager_data}()},
\code{\link{random_npi_generator}()}
}
\concept{mock}
