% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{remove_all_na}
\alias{remove_all_na}
\title{Remove columns and rows with all NAs}
\usage{
remove_all_na(x)
}
\arguments{
\item{x}{vector or \verb{<data.frame>}}
}
\value{
vector or \verb{<data.frame>} with columns and rows with all NAs removed
}
\description{
Remove columns and rows with all NAs
}
\keyword{internal}
