% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{roundup}
\alias{roundup}
\title{Round a numeric vector up to the nearest decimal place}
\usage{
roundup(x, digits = 2)
}
\arguments{
\item{x}{\verb{<dbl>} A numeric vector}

\item{digits}{\verb{<int>} Number of decimal places to round to}
}
\value{
a numeric vector rounded up to the nearest decimal place
}
\description{
Wrapper around \code{\link[janitor:round_half_up]{janitor::round_half_up()}}
}
