% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{search_in}
\alias{search_in}
\title{Search in data frame column if search term is not \code{NULL}}
\usage{
search_in(x, column, what)
}
\arguments{
\item{x}{vector or \verb{<data.frame>}}

\item{column}{\verb{<chr>} string of column name}

\item{what}{to search for in \code{column}}
}
\value{
vector or \verb{<data.frame>}
}
\description{
Search in data frame column if search term is not \code{NULL}
}
\keyword{internal}
