% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming.R
\name{search_in}
\alias{search_in}
\title{Search in data frame}
\usage{
search_in(df, dfcol, search)
}
\arguments{
\item{df}{A \verb{<data.frame>} or \verb{<tibble>}}

\item{dfcol}{A \verb{<character>} or \verb{<symbol>} specifying the column to search in}

\item{search}{A \verb{<character>} or \verb{<symbol>} specifying the search term}
}
\value{
A \verb{<data.frame>} or \verb{<tibble>}
}
\description{
Search in data frame
}
\examples{
x <- dplyr::tibble(y = 1:10,
                   z = letters[1:10])

search_in(df = x,
          dfcol = x$z,
          search = c("a", "j"))

}
