% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{sf_c}
\alias{sf_c}
\title{Concatenate Vectors}
\usage{
sf_c(...)
}
\arguments{
\item{...}{Any number of vectors, coerced to \verb{<chr>} vector, if necessary}
}
\value{
\verb{<chr>} concatenated vector
}
\description{
Concatenate Vectors
}
\keyword{internal}
