% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{sf_detect}
\alias{sf_detect}
\title{Detect by Regex}
\usage{
sf_detect(s, p)
}
\arguments{
\item{s}{\verb{<chr>} vector}

\item{p}{\verb{<chr>} regex pattern}
}
\value{
\verb{<lgl>} vector
}
\description{
Detect by Regex
}
\keyword{internal}
