% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringfish.R
\name{sf_extract}
\alias{sf_extract}
\title{Extract by Regex}
\usage{
sf_extract(s, p)
}
\arguments{
\item{s}{\verb{<character>} vector}

\item{p}{\verb{<character>} regex pattern}
}
\value{
\verb{<character>} vector
}
\description{
Extract by Regex
}
\examples{
sf_extract(sf_c(LETTERS, 0:9), "[A-Z]")

sf_extract(random_npi_generator(10), "[123]")

}
