% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{sf_ndetect}
\alias{sf_ndetect}
\title{Detect Opposite by Regex}
\usage{
sf_ndetect(s, p)
}
\arguments{
\item{s}{\verb{<chr>} vector}

\item{p}{\verb{<chr>} regex pattern}
}
\value{
\verb{<lgl>} vector
}
\description{
Detect Opposite by Regex
}
\keyword{internal}
