% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{sf_nextract}
\alias{sf_nextract}
\title{Extract Opposite by Regex}
\usage{
sf_nextract(s, p)
}
\arguments{
\item{s}{\verb{<chr>} vector}

\item{p}{\verb{<chr>} regex pattern}
}
\value{
\verb{<chr>} vector with pattern extracted
}
\description{
Extract Opposite by Regex
}
\keyword{internal}
