% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringfish.R
\name{sf_remove}
\alias{sf_remove}
\title{Remove by Regex}
\usage{
sf_remove(s, p)
}
\arguments{
\item{s}{\verb{<character>} vector}

\item{p}{\verb{<character>} regex pattern}
}
\value{
\verb{<character>} vector
}
\description{
Remove by Regex
}
\examples{
sf_remove(LETTERS, "A")

sf_remove(paste0(LETTERS, collapse = ""), "A")

}
