% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{sf_smush}
\alias{sf_smush}
\title{Collapse Vector}
\usage{
sf_smush(x, sep = "")
}
\arguments{
\item{x}{\verb{<chr>} vector}

\item{sep}{\verb{<chr>} separator; default is \code{""}}
}
\value{
\verb{<chr>} collapsed vector
}
\description{
Collapse Vector
}
\keyword{internal}
