% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{sf_strsplit}
\alias{sf_strsplit}
\title{Split string by delimiter}
\usage{
sf_strsplit(x, s, fix = TRUE)
}
\arguments{
\item{x}{\verb{<chr>} vector}

\item{s}{\verb{<chr>} delimiter to split by}

\item{fix}{\verb{<lgl>} fixed or regex; default is \code{TRUE}}
}
\value{
\verb{<list>} of split strings
}
\description{
Split string by delimiter
}
\keyword{internal}
