% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-helpers.R
\name{sf_sub}
\alias{sf_sub}
\title{Subset Vector by Range}
\usage{
sf_sub(x, start = 1, stop)
}
\arguments{
\item{x}{\verb{<chr>} vector}

\item{start}{\verb{<int>} index start; default is \code{1}}

\item{stop}{\verb{<int>} index end}
}
\value{
\verb{<chr>} vector
}
\description{
Subset Vector by Range
}
\keyword{internal}
