% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastverse.R
\name{strsort}
\alias{strsort}
\title{Parallel Sort wrapper using \code{\link[kit:psort]{kit::psort()}}}
\usage{
strsort(x)
}
\arguments{
\item{x}{\verb{<character>} vector. If other, will default to \code{\link[base:sort]{base::sort()}}}
}
\value{
\code{x} in sorted order
}
\description{
Parallel Sort wrapper using \code{\link[kit:psort]{kit::psort()}}
}
\examples{
strsort(random_npi_generator(10))

}
