% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastverse.R
\name{uniq_narm}
\alias{uniq_narm}
\title{Unique Values with NAs Removed}
\usage{
uniq_narm(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector
}
\description{
Unique Values with NAs Removed
}
\examples{
uniq_narm(c("A", NA, "A", 1))

}
