% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastverse.R
\name{uniq_vlen}
\alias{uniq_vlen}
\title{Unique Lengths of Vector}
\usage{
uniq_vlen(x)
}
\arguments{
\item{x}{vector}
}
\value{
\verb{<integer>} vector
}
\description{
Unique Lengths of Vector
}
\examples{
random_npi_generator(10) |> uniq_vlen()

}
